# docs/swagger.yaml
definitions:
...
  university_system_internal_domain_models.Course:
    properties:
      code:
        type: string
      created_at:
        type: string
      credits:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      teacher_id:
        type: integer
      updated_at:
        type: string
    type: object
  university_system_internal_domain_models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  university_system_internal_domain_models.Student:
    properties:
      birthdate:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      faculty:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      role:
        type: string
      student_year:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: n_rakhat@kbtu.kz
    name: Rakhat Nurtileu
    url: http://linkedin.com/in/nurtileu-rakhat-02aba3294/
  description: System для управления пользователями, курсами и ролями
  title: University System
  version: "1.0"
paths:
...
  /students:
    get:
      consumes:
      - application/json
      description: Возвращает список всех студентов в системе
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Student'
            type: array
        "401":
          description: Неавторизованный доступ
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить список всех студентов
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Добавляет нового студента в систему
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные нового студента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать нового студента
      tags:
      - students
  /students/{id}:
    delete:
      description: Удаляет студента по его ID
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID студента
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить студента
      tags:
      - students
    get:
      consumes:
      - application/json
      description: Возвращает данные студента по его ID
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID студента
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Student'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизованный доступ
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Студент не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить информацию о студенте
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Обновляет информацию о студенте
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID студента
        in: path
        name: id
        required: true
        type: string
      - description: Обновленные данные студента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить данные студента
      tags:
      - students
  /students/{id}/courses:
    get:
      consumes:
      - application/json
      description: Возвращает список курсов, на которые записан студент
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID студента
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Course'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить курсы студента
      tags:
      - students
  /students/{student_id}/courses/{course_id}:
    post:
      description: Записывает студента на указанный курс
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID студента
        in: path
        name: student_id
        required: true
        type: string
      - description: ID курса
        in: path
        name: course_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Записать студента на курс
      tags:
      - students
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"